#301 Move SqlGenerator interface along with SqlGeneratorBase to the API module.
