+++ pramen/core/src/main/scala/za/co/absa/pramen/core/sql/SqlGeneratorDb2.scala
@@ -16,6 +16,8 @@
 
 package za.co.absa.pramen.core.sql
 
+import za.co.absa.pramen.api.sql.{SqlColumnType, SqlConfig, SqlGeneratorBase}
+
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
